Updated logic for writing image version in sysfs
Removed target list and added check for presence of
sysfs entry 'select_image'

+++ rootdir/etc/init.qcom.post_boot.sh
@@ -1096,23 +1096,21 @@ case "$target" in
      ;;
 esac
 
-case "$target" in
-    "msm8226" | "msm8974" | "msm8610" | "apq8084" | "mpq8092" | "msm8610" | "msm8916" | "msm8994" | "msm8992")
-        # Let kernel know our image version/variant/crm_version
-        image_version="10:"
-        image_version+=`getprop ro.build.id`
-        image_version+=":"
-        image_version+=`getprop ro.build.version.incremental`
-        image_variant=`getprop ro.product.name`
-        image_variant+="-"
-        image_variant+=`getprop ro.build.type`
-        oem_version=`getprop ro.build.version.codename`
-        echo 10 > /sys/devices/soc0/select_image
-        echo $image_version > /sys/devices/soc0/image_version
-        echo $image_variant > /sys/devices/soc0/image_variant
-        echo $oem_version > /sys/devices/soc0/image_crm_version
-        ;;
-esac
+# Let kernel know our image version/variant/crm_version
+if [ -f /sys/devices/soc0/select_image ]; then
+    image_version="10:"
+    image_version+=`getprop ro.build.id`
+    image_version+=":"
+    image_version+=`getprop ro.build.version.incremental`
+    image_variant=`getprop ro.product.name`
+    image_variant+="-"
+    image_variant+=`getprop ro.build.type`
+    oem_version=`getprop ro.build.version.codename`
+    echo 10 > /sys/devices/soc0/select_image
+    echo $image_version > /sys/devices/soc0/image_version
+    echo $image_variant > /sys/devices/soc0/image_variant
+    echo $oem_version > /sys/devices/soc0/image_crm_version
+fi
 
 case "$target" in
     "msm8996")
